refactor : 자동차 이름 검증 역할을 InputView에서 Car로 이동
